+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/ComposerTextInputView.swift
@@ -7,7 +7,7 @@ import UIKit
 
 /// SwiftUI wrapper for a text field with multiple rows.
 struct ComposerTextInputView: UIViewRepresentable {
-
+    @EnvironmentObject var composerViewModel: MessageComposerViewModel
     @Injected(\.utils) private var utils
 
     @Binding var text: String
@@ -33,6 +33,9 @@ struct ComposerTextInputView: UIViewRepresentable {
         inputTextView.placeholderLabel.text = placeholder
         inputTextView.contentInsetAdjustmentBehavior = .never
         inputTextView.setContentCompressionResistancePriority(.streamLow, for: .horizontal)
+        inputTextView.onImagePasted = {
+            composerViewModel.imagePasted($0)
+        }
 
         if utils.messageListConfig.becomesFirstResponderOnOpen {
             inputTextView.becomeFirstResponder()